@@ -3,8 +3,8 @@ package cert
 import (
 	"crypto/tls"
 	"crypto/x509"
-	"io/ioutil"
 	"net/http"
+	"os"
 	"path/filepath"
 
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
@@ -15,13 +15,13 @@ func loadCertificates(caCertPool *x509.CertPool, certificatesDirPath string) err
 	if !fileutils.IsPathExists(certificatesDirPath, false) {
 		return nil
 	}
-	files, err := ioutil.ReadDir(certificatesDirPath)
+	files, err := os.ReadDir(certificatesDirPath)
 	err = errorutils.CheckError(err)
 	if err != nil {
 		return err
 	}
 	for _, file := range files {
-		caCert, err := ioutil.ReadFile(filepath.Join(certificatesDirPath, file.Name()))
+		caCert, err := os.ReadFile(filepath.Join(certificatesDirPath, file.Name()))
 		err = errorutils.CheckError(err)
 		if err != nil {
 			return err
