@@ -1,7 +1,7 @@
 package tests
 
 import (
-	"io/ioutil"
+	"os"
 	"path/filepath"
 	"testing"
 
@@ -31,7 +31,7 @@ func TestArtifactoryDownload(t *testing.T) {
 
 func flatDownload(t *testing.T) {
 	var err error
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -53,7 +53,7 @@ func flatDownload(t *testing.T) {
 		t.Error("Missing file c.tar.gz")
 	}
 
-	workingDir2, err := ioutil.TempDir("", "downloadTests")
+	workingDir2, err := os.MkdirTemp("", "downloadTests")
 	downloadTarget = workingDir2 + string(filepath.Separator)
 	if err != nil {
 		t.Error(err)
@@ -78,7 +78,7 @@ func flatDownload(t *testing.T) {
 func recursiveDownload(t *testing.T) {
 	uploadDummyFile(t)
 	var err error
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -100,7 +100,7 @@ func recursiveDownload(t *testing.T) {
 		t.Error("Missing file c.tar.gz")
 	}
 
-	workingDir2, err := ioutil.TempDir("", "downloadTests")
+	workingDir2, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -125,7 +125,7 @@ func recursiveDownload(t *testing.T) {
 func placeholderDownload(t *testing.T) {
 	uploadDummyFile(t)
 	var err error
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -147,7 +147,7 @@ func placeholderDownload(t *testing.T) {
 
 func includeDirsDownload(t *testing.T) {
 	var err error
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -170,7 +170,7 @@ func includeDirsDownload(t *testing.T) {
 }
 
 func exclusionsDownload(t *testing.T) {
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -195,7 +195,7 @@ func exclusionsDownload(t *testing.T) {
 }
 
 func explodeArchiveDownload(t *testing.T) {
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -244,7 +244,7 @@ func explodeDownloadAndVerify(t *testing.T, downloadParams *services.DownloadPar
 }
 
 func summaryDownload(t *testing.T) {
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
@@ -283,7 +283,7 @@ func summaryDownload(t *testing.T) {
 
 // Test downloading of two different files to the same path in the local machine. Only the first of them will be downloaded.
 func duplicateDownload(t *testing.T) {
-	workingDir, err := ioutil.TempDir("", "downloadTests")
+	workingDir, err := os.MkdirTemp("", "downloadTests")
 	if err != nil {
 		t.Error(err)
 	}
