@@ -7,7 +7,6 @@ import (
 	"errors"
 	"golang.org/x/exp/slices"
 	"io"
-	"io/ioutil"
 	"net/url"
 	"os"
 	"os/user"
@@ -174,7 +173,7 @@ func ListFilesByFilterFunc(path string, filterFunc func(filePath string) (bool, 
 		path += sep
 	}
 	var fileList []string
-	files, _ := ioutil.ReadDir(path)
+	files, _ := os.ReadDir(path)
 	path = strings.TrimPrefix(path, "."+sep)
 
 	for _, f := range files {
@@ -218,7 +217,7 @@ func ListFiles(path string, includeDirs bool) ([]string, error) {
 		path += sep
 	}
 	fileList := []string{}
-	files, _ := ioutil.ReadDir(path)
+	files, _ := os.ReadDir(path)
 	path = strings.TrimPrefix(path, "."+sep)
 
 	for _, f := range files {
@@ -345,7 +344,7 @@ func IsSshUrl(urlPath string) bool {
 }
 
 func ReadFile(filePath string) ([]byte, error) {
-	content, err := ioutil.ReadFile(filePath)
+	content, err := os.ReadFile(filePath)
 	err = errorutils.CheckError(err)
 	return content, err
 }
